Se puede jugar de forma local con dos jugadores

+++ server/game/game.h
@@ -17,7 +17,8 @@
 #include "manager/projectilemanager.h"
 #include "manager/explosionmanager.h"
 
-enum class RandomSelection {
+enum class RandomSelection
+{
     SpawnGun = 0,
     Explode,
     Nothing
@@ -32,7 +33,7 @@ private:
     bool initialize = true;
     std::vector<Map> maps;
     std::vector<uint8_t> player_ids;
-    std::map<uint8_t, Duck> ducks;
+    std::map<uint16_t, Duck> ducks;
     ExplosionManager explosions;
     ProjectileManager projectiles;
     std::map<uint8_t, uint8_t> victories;
